ahoy-components/src/main/java/za/co/lsd/ahoy/server/task/TaskExecutor.java
/*
 * Copyright  2022 LSD Information Technology (Pty) Ltd
 *
 *    Licensed under the Apache License, Version 2.0 (the "License");
 *    you may not use this file except in compliance with the License.
 *    You may obtain a copy of the License at
 *
 *        http://www.apache.org/licenses/LICENSE-2.0
 *
 *    Unless required by applicable law or agreed to in writing, software
 *    distributed under the License is distributed on an "AS IS" BASIS,
 *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *    See the License for the specific language governing permissions and
 *    limitations under the License.
 */

package za.co.lsd.ahoy.server.task;

import lombok.extern.slf4j.Slf4j;
import org.springframework.security.core.Authentication;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.stereotype.Component;

@Component
@Slf4j
public class TaskExecutor {
	private final TaskQueue taskQueue;

	public TaskExecutor(TaskQueue taskQueue) {
		this.taskQueue = taskQueue;
	}

	public void execute(Task task, TaskContext context) throws InterruptedException {
		Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
		if (authentication != null) {
			context.setAuthentication(authentication);
		}
		log.debug("Enqueuing task: {}, with context: {}", task.getName(), context);
		taskQueue.put(new TaskExecution(task, context));
	}
}
